Add repo-local copy of 'lfi'.
I'm pushed relucantly to do this.  I don't want to fight with `rebar`,
so I'm taking the lazy path.  If I used the scheme that I used in
the previous commit, where the lfi library is a separate dependency,
then `rebar` will change the paths depending on how Faulterl is used.

* top level source tree -> `deps/lfi`
* as a depdencency elsewhere in the `rebar`-managed source tree -> `../deps/lfi`

So, I've taken the fork of LFI that I've created and imported it under
the `priv/lfi` subdirectory.  That import is as of commit
79ff361dc8d624f6a040faddea41c85bebf38473 in the
`https://github.com/slfritchie/lfi` Git repo.